* master: (143 commits)
  update version info
  fix warnings: C4244: 'argument': conversion from 'intptr_t' to 'int'
  fix warnings: C4244: '=': conversion from 'uintptr_t' to 'DWORD'
  fix warnings: C4244: 'argument': conversion from 'intptr_t' to 'int'
  fix merge
  zlib-ng rev. 45a5149c
  include "GlobalUtils.hpp"
  GetFixedPaths
  WinSCPSecurity
  static wchar_t --> UnicodeString
  NextWord(): remove static buffer
  bool Result = false;
  nullptr
  TWinSCPFileSystem::Synchronize: rename param
  nullptr
  use std::unique_ptr<>
  use SAFE_DESTROY_EX
  cleanup
  W2MB()
  AnsiString::AnsiString(const wchar_t * Str, intptr_t Length, int CodePage)
  ...

# Conflicts:
#	src/NetBox/NetBox.cpp
#	src/NetBox/NetBox.rc

+++ src/core/SecureShell.cpp
@@ -2080,14 +2080,14 @@ bool TSecureShell::EventSelectLoop(uintptr_t MSec, bool ReadEventRequired,
       if (WSAIoctl(FSocket, SIO_IDEAL_SEND_BACKLOG_QUERY, NULL, 0, &BufferLen, sizeof(BufferLen), &OutLen, 0, 0) == 0)
       {
         DebugAssert(OutLen == sizeof(BufferLen));
-        if (FSendBuf < static_cast<int>(BufferLen))
+        if (FSendBuf < static_cast<intptr_t>(BufferLen))
         {
           LogEvent(FORMAT(L"Increasing send buffer from %d to %d", FSendBuf, static_cast<int>(BufferLen)));
           FSendBuf = BufferLen;
           setsockopt(FSocket, SOL_SOCKET, SO_SNDBUF, reinterpret_cast<const char *>(&BufferLen), sizeof(BufferLen));
         }
       }
-      FLastSendBufferUpdate = TicksAfter;
+      FLastSendBufferUpdate = (DWORD)TicksAfter;
     }
   }
   while (ReadEventRequired && (MSec > 0) && !Result);
